+++ packages/fx-core/src/component/questionV3.ts
@@ -1,7 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
-import * as path from "path";
-import * as fs from "fs-extra";
 import {
   CLIPlatforms,
   ContextV3,
@@ -96,6 +94,7 @@ import { manifestUtils } from "./resource/appManifest/utils";
 import { Constants } from "../plugins/resource/aad/constants";
 import { getQuestionsForDeployAPIM } from "./resource/apim";
 import { canAddSso } from "./feature/sso";
+import { getAddSPFxQuestionNode } from "./feature/spfx";
 
 export async function getQuestionsForProvisionV3(
   context: v2.Context,
@@ -495,32 +494,4 @@ export async function getNotificationTriggerQuestionNode(
   }
   res.condition = showNotificationTriggerCondition;
   return ok(res);
-}
-
-export async function getAddSPFxQuestionNode(
-  projectPath: string | undefined
-): Promise<Result<QTreeNode | undefined, FxError>> {
-  const spfx_add_feature = new QTreeNode({
-    type: "group",
-  });
-  spfx_add_feature.condition = { equals: TabSPFxNewUIItem.id };
-
-  const spfx_version_check = new QTreeNode(versionCheckQuestion);
-  spfx_add_feature.addChild(spfx_version_check);
-
-  if (projectPath) {
-    const yorcPath = path.join(projectPath, "SPFx", ".yo-rc.json");
-    if (await fs.pathExists(yorcPath)) {
-      const yorc = await fs.readJson(yorcPath);
-      const template = yorc["@microsoft/generator-sharepoint"]?.template;
-      if (template === undefined || template === "") {
-        const spfx_framework_type = new QTreeNode(frameworkQuestion);
-        spfx_version_check.addChild(spfx_framework_type);
-      }
-    }
-  }
-
-  const spfx_webpart_name = new QTreeNode(webpartNameQuestion);
-  spfx_version_check.addChild(spfx_webpart_name);
-  return ok(spfx_add_feature);
 }
